Extracting Profiler to a Separate Struct
